refactor: 의존성 주입을 관리하는 책임 분리
