+++ app/src/main/java/woowacourse/shopping/di/DependencyInjector.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.di
+
+import kotlin.reflect.KClass
+
+object DependencyInjector {
+    fun <T : Any> inject(targetClass: KClass<T>): T {
+        return DIContainer.getInstance(targetClass)
+    }
+}
