+++ app/src/main/java/woowacourse/shopping/di/RepositoryModule.kt
@@ -5,7 +5,7 @@ import woowacourse.shopping.data.DefaultProductRepository
 import woowacourse.shopping.domain.CartRepository
 import woowacourse.shopping.domain.ProductRepository
 
-class AppModule : DIModule {
+class RepositoryModule : DIModule {
     override fun register(container: DIContainer) {
         container.registerMapping(ProductRepository::class, DefaultProductRepository::class)
         container.registerMapping(CartRepository::class, DefaultCartRepository::class)
