+++ app/src/main/java/woowacourse/shopping/di/ViewModelFactory.kt
@@ -1,29 +1,10 @@
 package woowacourse.shopping.di
 
-import androidx.annotation.MainThread
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.ViewModelLazy
-import androidx.activity.ComponentActivity
-import androidx.lifecycle.viewmodel.CreationExtras
 
 class ViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return DIContainer.getInstance(modelClass.kotlin)
+        return DependencyInjector.inject(modelClass.kotlin)
     }
-}
-
-@MainThread
-inline fun <reified VM : ViewModel> ComponentActivity.provideViewModel(
-    noinline extrasProducer: (() -> CreationExtras)? = null,
-    noinline factoryProducer: (() -> ViewModelProvider.Factory)? = null,
-): Lazy<VM> {
-    val factoryPromise = factoryProducer ?: { ViewModelFactory() }
-
-    return ViewModelLazy(
-        VM::class,
-        { viewModelStore },
-        factoryPromise,
-        { extrasProducer?.invoke() ?: this.defaultViewModelCreationExtras },
-    )
 }
